rename SelectItem dir to SelectItemFormField and mv SelectItemFormField component inside grouping folder
