+++ src/components/react-hook-form/SelectItemFormField/__test__/replacePlaceholderColorWithTextColor.test.ts
