+++ src/components/react-hook-form/SelectItemFormField/index.ts
@@ -1,2 +1,3 @@
+export * from './SelectItemFormField'
 export * from './SelectItemFormFieldDialog'
 export * from './SelectItemFormFieldInput'
