+++ src/components/react-hook-form/SelectItemFormField/replacePlaceholderColorWithTextColor.ts
