+++ src/components/react-hook-form/index.ts
@@ -16,5 +16,4 @@ export * from './RadioFormField'
 export * from './SearchFormField'
 export * from './useHandleSubmit'
 export * from './SelectItemFormField'
-export * from './SelectItem'
 export * from './util'
